feat: refactor: create post endpoint

+++ Reaction.Presentation/Controllers/CommentReactionController.cs
@@ -1,5 +1,4 @@
 ﻿using AutoMapper;
-using InfinityNetServer.BuildingBlocks.Application.Contracts;
 using InfinityNetServer.BuildingBlocks.Application.DTOs.Responses.File;
 using InfinityNetServer.BuildingBlocks.Application.DTOs.Responses.Profile;
 using InfinityNetServer.BuildingBlocks.Application.Exceptions;
@@ -35,8 +34,6 @@ namespace InfinityNetServer.Services.Reaction.Presentation.Controllers
         IMapper mapper,
         IStringLocalizer<ReactionSharedResource> localizer,
         CommonProfileClient profileClient,
-        CommonCommentClient commentClient,
-        IMessageBus messageBus,
         CommonFileClient fileClient,
         ICommentReactionService service) : BaseApiController(authenticatedUserService)
     {
@@ -51,8 +48,7 @@ namespace InfinityNetServer.Services.Reaction.Presentation.Controllers
             [FromQuery] int limit = 10,
             [FromQuery] string type = "Like")
         {
-            Guid currentProfileId = GetCurrentProfileId != null ? GetCurrentProfileId().Value
-                : throw new BaseException(BaseError.PROFILE_NOT_FOUND, StatusCodes.Status404NotFound);
+            Guid currentProfileId = GetCurrentProfileId();
 
             ReactionType reactionType = Enum.Parse<ReactionType>(type);
 
@@ -91,8 +87,11 @@ namespace InfinityNetServer.Services.Reaction.Presentation.Controllers
                     // Process Owner
                     if (profileDict.TryGetValue(commentItem.ProfileId, out var ownerProfile))
                     {
-                        var avatar = photoMetadataDict.GetValueOrDefault(ownerProfile.Avatar.Id);
-                        ownerProfile.Avatar = avatar;
+                        if (ownerProfile.Avatar != null)
+                        {
+                            var avatar = photoMetadataDict.GetValueOrDefault(ownerProfile.Avatar.Id);
+                            ownerProfile.Avatar = avatar;
+                        }
                         commentReactionResponse.Profile = ownerProfile;
                     }
 
@@ -113,15 +112,14 @@ namespace InfinityNetServer.Services.Reaction.Presentation.Controllers
         {
             try
             {
-                Guid currentProfileId = GetCurrentProfileId != null ? GetCurrentProfileId().Value
-                    : throw new BaseException(BaseError.PROFILE_NOT_FOUND, StatusCodes.Status404NotFound);
+                Guid currentProfileId = GetCurrentProfileId();
 
                 var reaction = await service.Save(new CommentReaction
                 {
                     CommentId = Guid.Parse(commentId),
                     ProfileId = currentProfileId,
                     Type = Enum.Parse<ReactionType>(request.Reaction)
-                }, commentClient, messageBus);
+                });
 
                 var reactionCounts = await service.CountByCommentId([commentId]);
 
@@ -146,8 +144,12 @@ namespace InfinityNetServer.Services.Reaction.Presentation.Controllers
         {
             try
             {
-                Guid currentProfileId = GetCurrentProfileId != null ? GetCurrentProfileId().Value
-                    : throw new BaseException(BaseError.PROFILE_NOT_FOUND, StatusCodes.Status404NotFound);
+                Guid currentProfileId = GetCurrentProfileId();
+
+                var existedReaction = await service.GetByCommentIdAndProfileId(commentId, currentProfileId.ToString());
+
+                if(!IsOwner(existedReaction.ProfileId.ToString()))
+                    throw new BaseException(BaseError.NOT_HAVE_PERMISSION, StatusCodes.Status403Forbidden);
 
                 var reaction = await service.Delete(commentId, currentProfileId.ToString());
 
@@ -155,7 +157,7 @@ namespace InfinityNetServer.Services.Reaction.Presentation.Controllers
 
                 return Ok(new
                 {
-                    ReactionCounts = reactionCounts[0].countDetails
+                    ReactionCounts = reactionCounts.Count > 0 ? reactionCounts[0].countDetails : new Dictionary<ReactionType, int>()
                 });
             }
             catch (Exception ex)
